refactor: organize i18n by layer
