+++ web/src/infra/i18n/index.ts
@@ -2,6 +2,6 @@ import { nextI18Next } from './config';
 
 export * from './config';
 export * from './hooks';
-export * from './router';
+export * from './utils';
 
 export const { Link } = nextI18Next;
