--- web/src/infra/i18n/router/index.ts
@@ -1,1 +1,0 @@
-export * from './router';
