+++ web/src/infra/i18n/types/index.ts
