+++ web/src/infra/i18n/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './href';
+export * from './router';
